ci: build on ubuntu with GCC
# Motivation

To properly support ubuntu 24.04, building with `make` is not
sufficient. Just managing the dependencies is difficult. Since this
repository is already using `cmake` and `vcpkg` for building on windows,
this commit is simply extending this for ubuntu-24.04.

To build locally, you simply need to run
`python continuous-integration/build.py` when you have all of the
dependencies installed. These you can install by running
`continuous-integration/ubuntu-24.04/setup.sh`.

Alternatively you can just do the same as the Github Actions workflow
run and do the same as it does.

The usage of `ninja` as generator and `gcc` as compiler is completely
arbitrary. However the setup is configured so that adding support for a
different compiler such as `clang` could be done with minimal effort.

Since we now use `CMakePresets.json` to provide `vcpkg` to `CMake`, this
forces a bump to `CMake` version 3.19.

# Rejected alternatives

The system dependencies that we rely on could be installed with `nix` +
`devenv`, but I decided that this might be a bit too
controversial/overkill for the initial ubuntu build.

Using `AppVeyor` to build targeting different platforms + compilers
could be an option, however this is difficult for anyone that is not the
repository owner to contribute, as the `AppVeyor` settings are not
something that are visible to others.

By using `Github Actions` makes it possible for others that fork this
repository to have working builds from the very start.

# Future work

This commit only ensures that the build actually succeeds, if the
resulting artifacts can actually be run is a different task altogether.

There are a lot of compiler warnings like
```
warning: invalid use of incomplete type 'struct CommonGameState'
```

That is printed to stdout when building, but these would need to be
addressed in a future commit.